+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 Version 0.5
 -----------
 
+* Added support for column families.
+* Added SstFileWriter
+* Tested with RocksDB 4.6.1
 
 Version 0.4
 -----------